services/speech-to-text/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch==2.3.1
openai-whisper==2024.1.1
